Added led timer

+++ Final/mainproject/h/counter.h
@@ -24,7 +24,7 @@ public:
 
     static void setTim16(uint16_t p);
 
-    static void setTim17(uint16_t p);
+    static void setTim17(uint32_t p);
 
     static void incrementTime(int &seconds);
